+++ lib/Reconciler.spec.lua
@@ -1,8 +1,8 @@
 return function()
 	local Reconciler = require(script.Parent.Reconciler)
 
-	it("should reify booleans as nil", function()
-		local booleanReified = Reconciler.reify(false)
+	it("should mount booleans as nil", function()
+		local booleanReified = Reconciler.mount(false)
 		expect(booleanReified).to.never.be.ok()
 	end)
 end